+++ hbw/app/javascript/packs/hbw/components/form/services_table/table_title.tsx
@@ -4,6 +4,8 @@ import {
   withCallbacks
 } from 'shared/hoc';
 import TranslationContext, { ContextType as TranslationContextType } from 'shared/context/translation';
+import Select, { MultiValue } from 'react-select';
+import { Column } from './services_table';
 
 type Props = {
   name: string,
@@ -11,6 +13,11 @@ type Props = {
   search: {
     hidden: boolean,
     onSearchQueryChange: React.ChangeEventHandler<HTMLInputElement>
+  },
+  columnsControl: {
+    visibleColumns: Column[],
+    columns: Column[],
+    onChange: ((newValue: MultiValue<Column>) => void)
   },
   task: {
     key: string,
@@ -33,6 +40,26 @@ const HBWServicesTableTitle: React.FC<Props> = (props) => {
             type="search"
             className="form-control"
             onChange={props.search.onSearchQueryChange}
+          />
+        </div>
+      </div>
+      <div className="right">
+        <div className="columns-setting">
+          <Select
+            name={props.name}
+            className="react-select-container"
+            classNamePrefix="react-select"
+            isMulti
+            closeMenuOnSelect={false}
+            hideSelectedOptions={false}
+            controlShouldRenderValue={false}
+            isClearable={false}
+            isSearchable={false}
+            placeholder={translate('components.services_table.table.title.columns_setting.placeholder')}
+            getOptionValue={option => option.name }
+            defaultValue={props.columnsControl.visibleColumns}
+            onChange={props.columnsControl.onChange}
+            options={props.columnsControl.columns}
           />
         </div>
       </div>